// src/api/models/createAssessmentReqDTOMajorField.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

/**
 * 전공 계열. 전공 무관이면 null
 * @nullable
 */
export type CreateAssessmentReqDTOMajorField = typeof CreateAssessmentReqDTOMajorField[keyof typeof CreateAssessmentReqDTOMajorField] | null;


export const CreateAssessmentReqDTOMajorField = {
  HUMANITIES_SOCIAL: 'HUMANITIES_SOCIAL',
  NATURAL_SCIENCE: 'NATURAL_SCIENCE',
  LANGUAGE: 'LANGUAGE',
  BUSINESS: 'BUSINESS',
  ECONOMICS: 'ECONOMICS',
  MEDIA_COMMUNICATION: 'MEDIA_COMMUNICATION',
  MATH_STATISTICS: 'MATH_STATISTICS',
  COMPUTER_SCIENCE: 'COMPUTER_SCIENCE',
  ENGINEERING: 'ENGINEERING',
  ART_DESIGN: 'ART_DESIGN',
} as const;